requirements.txt
discord.py
google-generativeai
python-dotenv
firebase-admin
Pillow
Flask